Experimenting with a simplified mixing method
This is a standard delay that attempts to simplify the way that feedback and dry signal is mixed and hopefully improve audio quality. Also reduces the resolution of the delay control to 8 bit to maybe help with noise. A capacitor might be needed on the delay time pin to reduce the functional impedance.